Don’t override server-supplied timeouts
The current implementation of PromptBuilder makes a somewhat naive assumption that the provided `timeout` setting should apply to three different kinds of time-outs.  The intent seems to apply best to the `initial-timeout`, or the amount of time the input should wait for the user to start signaling a response. But it’s also default-applied to `max-silence` and `inter-digit` timeouts. This is bad for callers who may take up to 4 seconds to start talking, and then have to wait up to 4 additional seconds after they stop talking for an answer. The server-supplied timeouts (LumenVox is around 800ms) should be the default, while still being possible to override them by the caller to `#ask`.